Sidecar Rev B clock config
This diff splits the clock generator config payload into different files for
different board revs. An rev B config is provided which sets the outputs
according to board changes made.
